AI-engine/requirements.txt
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
chromadb
transformers
sentence_transformers
torch
accelerate
bitsandbytes